+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 3.8.1 under development
 
-- Enh #230: Add support `BackedEnum` value to `Html::addCssClass()` (@terabytesoftw)
+- Enh #230: Add support `BackedEnum` value to `Html::addCssClass()`, `Tag::addClass()` and `Tag::class()` methods (@terabytesoftw)
 
 ## 3.8.0 October 29, 2024