+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Yii HTML Change Log
 
-## 3.8.1 under development
+## 3.9.0 November 29, 2024
 
 - Enh #230: Add backed enumeration value support to `Html::addCssClass()`, `Tag::addClass()` and `Tag::class()`
   methods (@terabytesoftw)